Move process event call from case to action
