+++ src/main/java/uk/gov/ons/ctp/response/collection/exercise/CollectionExerciseApplication.java
@@ -94,17 +94,6 @@ public class CollectionExerciseApplication {
   }
 
   /**
-   * The RestUtility bean for the Case service
-   *
-   * @return the RestUtility bean for the Case service
-   */
-  @Bean
-  @Qualifier("caseRestUtility")
-  public RestUtility caseRestUtility() {
-    return new RestUtility(appConfig.getCaseSvc().getConnectionConfig());
-  }
-
-  /**
    * The RestUtility bean for the CollectionInstrument service
    *
    * @return the RestUtility bean for the CollectionInstrument service
