+++ src/main/java/uk/gov/ons/ctp/response/collection/exercise/config/ActionSvc.java
@@ -14,5 +14,6 @@ public class ActionSvc {
   private String actionRulePath;
   private String actionRulesForActionPlanPath;
   private String actionPlanPath;
+  private String processEventPath;
   private boolean deprecated;
 }
