--- src/main/java/uk/gov/ons/ctp/response/collection/exercise/config/CaseSvc.java
@@ -1,13 +1,0 @@
-package uk.gov.ons.ctp.response.collection.exercise.config;
-
-import lombok.Data;
-import net.sourceforge.cobertura.CoverageIgnore;
-import uk.gov.ons.ctp.response.collection.exercise.lib.common.rest.RestUtilityConfig;
-
-/** Application Config bean for the connection details to the Action Service */
-@Data
-@CoverageIgnore
-public class CaseSvc {
-  private RestUtilityConfig connectionConfig;
-  private String processEventPath;
-}
