+++ src/test/java/uk/gov/ons/ctp/response/collection/exercise/client/ActionSvcClientTest.java
@@ -1,6 +1,7 @@
 package uk.gov.ons.ctp.response.collection.exercise.client;
 
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
 import static org.mockito.Matchers.any;
 import static org.mockito.Matchers.eq;
 import static org.mockito.Mockito.times;
@@ -38,9 +39,11 @@ import uk.gov.ons.ctp.response.collection.exercise.lib.common.rest.RestUtilityCo
 @RunWith(MockitoJUnitRunner.class)
 public class ActionSvcClientTest {
 
+  private static final String EVENT_TAG = "mps";
   private static final String ACTION_PLAN_NAME = "APName";
   private static final String ACTION_PLAN_DESCRIPTION = "APDescription";
   private static final String COLLECTION_EXERCISE_ID = "14fb3e68-4dca-46db-bf49-04b84e07e77c";
+  private static final UUID COLLECTION_EXERCISE_UUID = UUID.fromString(COLLECTION_EXERCISE_ID);
   private static final String ACTION_PLAN_ID = "14fb3e68-5dca-46db-bf49-04b84e07e77c";
 
   @Spy private RestUtility restUtility = new RestUtility(RestUtilityConfig.builder().build());
@@ -198,5 +201,46 @@ public class ActionSvcClientTest {
             HttpMethod.GET,
             null,
             new ParameterizedTypeReference<List<ActionPlanDTO>>() {});
+  }
+
+  @Test
+  public void processEvent_204Response() {
+    // Given
+    ResponseEntity responseEntity = new ResponseEntity<>(HttpStatus.NO_CONTENT);
+    when(restTemplate.postForEntity(any(URI.class), any(HttpEntity.class), any()))
+        .thenReturn(responseEntity);
+
+    // When
+    boolean result = actionSvcRestClient.processEvent(EVENT_TAG, COLLECTION_EXERCISE_UUID);
+
+    // Then
+    verify(restTemplate, times(1)).postForEntity(any(URI.class), any(HttpEntity.class), any(null));
+    assertTrue(result);
+  }
+
+  @Test(expected = HttpClientErrorException.class)
+  public void processEvent_400Response() {
+    // Given
+    when(restTemplate.postForEntity(any(URI.class), any(HttpEntity.class), any()))
+        .thenThrow(new HttpClientErrorException(HttpStatus.BAD_REQUEST));
+
+    // When
+    boolean result = actionSvcRestClient.processEvent(EVENT_TAG, COLLECTION_EXERCISE_UUID);
+
+    // Then
+    verify(restTemplate, times(1)).postForEntity(any(URI.class), any(HttpEntity.class), any(null));
+  }
+
+  @Test(expected = HttpClientErrorException.class)
+  public void processEvent_500Response() {
+    // Given
+    when(restTemplate.postForEntity(any(URI.class), any(HttpEntity.class), any()))
+        .thenThrow(new HttpClientErrorException(HttpStatus.INTERNAL_SERVER_ERROR));
+
+    // When
+    actionSvcRestClient.processEvent(EVENT_TAG, COLLECTION_EXERCISE_UUID);
+
+    // Then
+    verify(restTemplate, times(1)).postForEntity(any(URI.class), any(HttpEntity.class), any(null));
   }
 }
