+++ src/test/java/uk/gov/ons/ctp/response/collection/exercise/service/EventServiceTest.java
@@ -25,7 +25,7 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.Spy;
 import org.mockito.runners.MockitoJUnitRunner;
-import uk.gov.ons.ctp.response.collection.exercise.client.CaseSvcClient;
+import uk.gov.ons.ctp.response.collection.exercise.client.ActionSvcClient;
 import uk.gov.ons.ctp.response.collection.exercise.client.SurveySvcClient;
 import uk.gov.ons.ctp.response.collection.exercise.config.ActionSvc;
 import uk.gov.ons.ctp.response.collection.exercise.config.AppConfig;
@@ -51,7 +51,7 @@ public class EventServiceTest {
   private static final UUID COLLEX_UUID = UUID.fromString("f03206ee-137d-41e3-af5c-2dea393bb360");
   private static final int EXERCISE_PK = 6433;
 
-  @Mock private CaseSvcClient caseSvcClient;
+  @Mock private ActionSvcClient actionSvcClient;
 
   @Mock private SurveySvcClient surveySvcClient;
 
@@ -593,7 +593,7 @@ public class EventServiceTest {
 
     // Then
     verify(eventRepository, atMost(1)).findByStatus(EventDTO.Status.SCHEDULED);
-    verify(caseSvcClient, never()).processEvent(any(), any());
+    verify(actionSvcClient, never()).processEvent(any(), any());
   }
 
   @Test
@@ -613,7 +613,7 @@ public class EventServiceTest {
 
     // Then
     verify(eventRepository, atMost(1)).findByStatus(EventDTO.Status.SCHEDULED);
-    verify(caseSvcClient, never()).processEvent(any(), any());
+    verify(actionSvcClient, never()).processEvent(any(), any());
   }
 
   @Test
@@ -633,7 +633,7 @@ public class EventServiceTest {
 
     // Then
     verify(eventRepository, atMost(1)).findByStatus(EventDTO.Status.SCHEDULED);
-    verify(caseSvcClient, never()).processEvent(any(), any());
+    verify(actionSvcClient, never()).processEvent(any(), any());
     try {
       verify(collectionExerciseService, atMost(1))
           .transitionCollectionExercise(
